refactor: Comment 검증 로직
User 체크를 조금 더 명시적이게 변경

+++ src/main/java/techcourse/myblog/service/CommentService.java
@@ -9,6 +9,7 @@ import techcourse.myblog.domain.User;
 import techcourse.myblog.service.dto.CommentRequest;
 import techcourse.myblog.service.dto.CommentResponse;
 import techcourse.myblog.service.exception.InvalidAuthorException;
+import techcourse.myblog.service.exception.NoCommentException;
 
 import javax.transaction.Transactional;
 import java.util.List;
@@ -37,17 +38,32 @@ public class CommentService {
 
     @Transactional
     public Comment update(CommentRequest commentRequest, User user, Long commentId) {
-        Comment comment = findByCommenterAndId(user, commentId);
+        Comment comment = findComment(user, commentId);
         return comment.updateContents(commentRequest.getContents());
     }
 
-    public Comment findByCommenterAndId(User user, Long commentId) {
-        return commentRepository.findByCommenterAndId(user, commentId)
-                .orElseThrow(() -> new InvalidAuthorException("작성자가 아닙니다"));
+    public void deleteById(Long commentId, User user) {
+        checkAuthor(user, commentId);
+        commentRepository.deleteById(commentId);
     }
 
-    public void deleteById(Long commentId, User user) {
-        findByCommenterAndId(user, commentId);
-        commentRepository.deleteById(commentId);
+    public Comment findComment(User user, Long commentId) {
+        Comment comment = findById(commentId);
+        if (comment.isCommentor(user)) {
+            throw new InvalidAuthorException("작성자가 아닙니다");
+        }
+        return comment;
+    }
+
+    private Comment findById(Long commentId) {
+        return commentRepository.findById(commentId)
+                .orElseThrow(() -> new NoCommentException("댓글이 존재하지 않습니다"));
+    }
+
+    private void checkAuthor(User user, Long commentId) {
+        Comment comment = findById(commentId);
+        if (comment.isCommentor(user)) {
+            throw new InvalidAuthorException("작성자가 아닙니다");
+        }
     }
 }
